feat: add generic type to `TransactionManager`
This commit enhances the `TransactionManager` abstract class by introducing
a generic type parameter. This allows for greater flexibility and
type safety when working with different types of transactions.

Specifically, the `TransactionManager<T = any>` now accepts a type parameter `T`.
This allows subclasses to specify the type of data they will be
managing within the transaction. The default type is set to `any` to
maintain backward compatibility.